project/requirements.txt
Django==4.1
django_filter==22.1
djangorestframework==3.13.1
drf_spectacular==0.23.1
pytest==7.1.2
python-decouple==3.6
